remove old validate code
