+++ packages/perseus/src/widgets/numeric-input/numeric-input-validator.test.ts
@@ -1,3 +1,30 @@
-function numericInputValidator() {}
+import {testDependencies} from "../../../../../testing/test-dependencies";
+import * as Dependencies from "../../dependencies";
 
-export default numericInputValidator;
+import {maybeParsePercentInput} from "./numeric-input-validator";
+
+describe("maybeParsePercentInput utility function", () => {
+    beforeEach(() => {
+        jest.spyOn(Dependencies, "getDependencies").mockReturnValue(
+            testDependencies,
+        );
+    });
+
+    it("returns a percent if the input is a percent", () => {
+        const result = maybeParsePercentInput("3%", false);
+
+        expect(result).toBe(3);
+    });
+
+    it("returns the input if its not valid", () => {
+        const result = maybeParsePercentInput("asdasd%", false);
+
+        expect(result).toBe("asdasd%");
+    });
+
+    it("normalizes values", () => {
+        const result = maybeParsePercentInput("20%", true);
+
+        expect(result).toBe(0.2);
+    });
+});
